Implement `ErodeImage` module support; move all modules-related logic to a separate `hcs_modules_factory`
